# examples/customer-credit-rating/my-credit-rating-agency/src/main/resources/META-INF/openapi.yaml
---
openapi: 3.0.2
info:
  title: ACME credit rating agency
  version: 1.0.0
  description: Custom Credit Rating Agency API
paths:
  /creditratings/{tin}:
    summary: Path used to manage a single CreditRating.
    description: "The REST endpoint/path used to get, update, and delete single instances\
      \ of an `CreditRating`.  This path contains `GET`, `PUT`, and `DELETE` operations\
      \ used to perform the get, update, and delete tasks, respectively."
    get:
      parameters:
        - examples:
            dynamic:
              value: "ABCD-1234-X"
          name: tin
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditRating'
              examples:
                dynamic:
                  value: |-
                    {
                      "tin": "{{ request.path[2] }}",
                      "score": "{{ randomInt(100) }}"
                    }
          description: Successful response - returns a single `CreditRating`.
      operationId: getCreditRating
      summary: Get a CreditRating
      description: Gets the details of a single instance of a `CreditRating`.
    put:
      parameters:
        - examples:
            score:
              value: "ABCD-1234-X"
          name: tin
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Updated `CreditRating` information.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditRating'
            examples:
              score:
                value:
                  score: 71
                  tin: ABCD-1234-X
        required: true
      responses:
        "202":
          description: Successful response.
          content:
            application/json:
              schema:
                type: string
              examples:
                score:
                  value: ""
      operationId: updateCreditRating
      summary: Update a CreditRating
      description: Updates an existing `CreditRating`.
    delete:
      parameters:
        - examples:
            deleted:
              value: "ABCD-1234-X"
          name: tin
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Successful response
          content:
            application/json:
              schema:
                type: string
              examples:
                deleted:
                  value: ""
      operationId: deleteCreditRating
      summary: Delete a CreditRating
      description: Deletes an existing `CreditRating`.
    parameters:
      - examples:
          tin:
            value: ABCD-1234-X
        name: tin
        in: path
        required: true
        schema:
          type: string
  /creditratings:
    summary: Path used to manage the list of credit ratings.
    description: "The REST endpoint/path used to list and create zero or more `CreditRating`\
      \ entities.  This path contains a `GET` and `POST` operation to perform the\
      \ list and create tasks, respectively."
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreditRating'
              examples:
                dynamic:
                  value: |-
                    [
                      {
                        "tin": "{{ randomString(20) }}",
                        "score": "{{ randomInt(100) }}"
                      },
                      {
                        "tin": "{{ randomString(20) }}",
                        "score": "{{ randomInt(100) }}"
                      }
                    ]
          description: Successful response - returns an array of `CreditRating` entities.
      operationId: getCreditRatings
      summary: List All CreditRatings
      description: Gets a list of all `CreditRating` entities.
    post:
      requestBody:
        description: A new `CreditRating` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditRating'
            examples:
              score:
                value:
                  score: 87
                  tin: ABCD-1234-X
        required: true
      responses:
        "201":
          description: Successful response.
          content:
            application/json:
              schema:
                type: string
              examples:
                created:
                  value: ""
      operationId: createCreditRating
      summary: Create a CreditRating
      description: Creates a new instance of a `CreditRating`.
components:
  schemas:
    CreditRating:
      title: Root Type for SubmitCreditRatingRequest
      description: ""
      required:
        - score
        - tin
      type: object
      properties:
        score:
          description: Credit Rating Score
          type: integer
        tin:
          description: Tax Identification Number
          type: string
      example:
        tin: ABC123899D
        score: 50